Fix drawerLayout swipe onBackPress

+++ app/src/main/java/com/imadev/foody/ui/MainActivity.kt
@@ -107,7 +107,7 @@ class MainActivity : AppCompatActivity() {
                 endId: Int,
                 progress: Float
             ) {
-                isDrawerActive = progress.toInt() > 0
+                isDrawerActive = progress > 0
             }
 
             override fun onTransitionCompleted(motionLayout: MotionLayout?, currentId: Int) {
